[MCUX-80525][MCXE31]Update C startup for IAR adding __iar_init$$done.
- This patch fixes some source are not relocated by "the initialize by copy" directive in linker.
Sections that are needed for initialization are not affected by the initialize by copy directive. This includes the __low_level_init function and anything it references.
Anything reachable from the program entry label is considered needed for initialization unless reached via a section fragment with a label starting with __iar_init$$done.
But __iar_init$done can't be used in C file, so changed to customized init and startup entry.

Signed-off-by: Chay Guo <changyi.guo@nxp.com>