[MCUX-83178][MCXL20] Fix Aon FRO setup function in clock driver
CLOCK_SetupFROAonClocking function disables AON FRO which is not selected
as Aon Root Clock by this function. In case of 10M FRO it can interfere
with 10M usage in main domain as LPIRC main clock.
Change behavior of the function such that it does not disable 10M FRO
when 2M FRO is selected as AON Root Clock.

Signed-off-by: Michal Smola <michal.smola@nxp.com>